apps/api/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pandas>=2.2.0
numpy>=1.26.0
yfinance>=0.2.40
requests>=2.31.0
pydantic>=2.7.0
pydantic-settings>=2.3.0
sqlmodel>=0.0.21
python-dotenv>=1.0.1
